Add files via upload
This is my Python Project. In this project I created a simple math quiz. I name it as an Brain Tease. Because it contains Addition, Subtraction, Multiplication Questions, And it shows the how many correct answer and how many wrong answer was given by user and how much he/she taken to the complete that quiz.